+++ backend/internal/config/new.go
@@ -29,13 +29,6 @@ type User struct {
 	Port string `yaml:"port"`
 }
 
-type Twitter struct {
-	Apikey       string `yaml:"api_key"`
-	ApiKeySecret string `yaml:"api_key_secret"`
-	BearerToken  string `yaml:"bearer_token"`
-	UserId       string `yaml:"user_id"`
-}
-
 type AuthToken struct {
 	SecretKey string        `yaml:"secret_key"`
 	Expires   time.Duration `yaml:"expires"`
